Components

EPHS.Components — Module

The Components module defines the primitive systems: StorageComponents, ReversibleComponents, and IrreversibleComponents.

EPHS.Components.CVar — Type

Constraint multiplier variable

EPHS.Components.Component — Type

Component is a subtype of EPHS.AbstractSystems.AbstractSystem. Concrete subtypes of Component are StorageComponent, ReversibleComponent, and IrreversibleComponent.

EPHS.Components.Constraint — Type

Constraint(residual::SymExpr)

The residual is forced to be zero by the corresponding constraint variable.

EPHS.Components.EffortPort — Type

EffortPort(quantity::Quantity, effort::SymExpr)

A EffortPort of a ReversibleComponent used to define a transformer-like coupling (in combination with a corresponding FlowPort).

EPHS.Components.FlowPort — Type

FlowPort(quantity::Quantity, flow::SymExpr)

A FlowPort of a ReversibleComponent used to define a gyrator-like coupling or a transformer-like coupling (in combination with a corresponding EffortPort).

EPHS.Components.IrreversibleComponent — Type

IrreversibleComponent(ports::Dtry{IrreversiblePort})

An IrreversibleComponent is a primitive system representing irreversible processes.

Fields

ports: directory of IrreversiblePorts

EPHS.Components.IrreversiblePort — Type

IrreversiblePort(quantity::Quantity, flow::SymExpr)

A port of an IrreversibleComponent provides the flow variable.

Fields

quantity: EPHS.AbstractSystems.Quantity of the port

flow: EPHS.SymbolicExpressions.SymExpr defining the flow variable

EPHS.Components.Par — Type

Par(box_path::DtryPath, par_path::DtryPath, value::Float64

Symbolic parameter identified by box_path * par_path.

Fields

box_path: indicates the subsystem to which the parameter belongs

par_path: distinguishes between different parameters of the same subsystem

value: (default) value of the parameter

EPHS.Components.Par — Method

Par(name::Symbol, value::Float64

Constructs a symbolic Parameter with box_path=DtryPath() and par_path=DtryPath(name).

Arguments

name: name of the parameter

value: (default) value of the parameter

EPHS.Components.ReversibleComponent — Type

ReversibleComponent(ports::Dtry{ReversiblePort})

A ReversibleComponent is a primitive system representing reversible dynamics, transformations, or constraints.

Fields

ports: directory of ReversiblePorts

EPHS.Components.ReversiblePort — Type

ReversiblePort(variant::Union{FlowPort,EffortPort,StatePort,Constraint})

A 'port' of a ReversibleComponent can be a

FlowPort which provides a flow variable

EffortPort which provides an effort variable

StatePort which consumes a state variable

Constraint which defines a residual and a constraint variable

EPHS.Components.StatePort — Type

StatePort(quantity::Quantity)

A StatePort of a ReversibleComponent needed if a gyrator/transformer-like coupling or a constraint depends on a state variable of a system which is not already connected via a FlowPort or an EffortPort.

EPHS.Components.StorageComponent — Type

StorageComponent(ports::Dtry{StoragePort}, energy::SymExpr)

A StorageComponent is a primitive system representing energy storage.

Fields

ports: directory of StoragePorts

energy: EPHS.SymbolicExpressions.SymExpr defining the energy

A StorageComponent should be constructed using the function StorageComponent(ports::Dtry{Quantity}, energy::SymExpr), which uses symbolic differentiation to compute the effort variables of the ports.

EPHS.Components.StorageComponent — Method

StorageComponent(ports::Dtry{Quantity}, energy::SymExpr)

A StorageComponent is a primitive system representing energy storage.

Arguments

ports: directory of Quantitys defining the interface of the component (power ports only)

energy: EPHS.SymbolicExpressions.SymExpr defining the energy

EPHS.Components.StoragePort — Type

StoragePort(quantity::Quantity, effort::SymExpr)

A port of a StorageComponent provides the effort variable.

Fields

quantity: EPHS.AbstractSystems.Quantity of the port

effort: EPHS.SymbolicExpressions.SymExpr defining the effort variable

EPHS.Components.θ₀ — Constant

Symbolic parameter, see Par, representing the exergy environment temperature with a (default) value of 300K.

EPHS.Components.π₀ — Constant

Symbolic parameter, see Par, representing the exergy environment pressuere with a (default) value of 10⁵Pa.